+++ src/frame.cpp
@@ -752,6 +752,20 @@ bool frame::vi_exec_command(wex::ex_command& command)
         page_prev();
         handled = true;
       }
+      else if (cmd.starts_with(":tabdo"))
+      {
+        const auto& arg(wex::find_after(cmd, " "));
+
+        for (int page = m_editors->GetPageCount() - 1; page >= 0; page--)
+        {
+          if (auto* stc = (wex::stc*)m_editors->GetPage(page);
+              stc != nullptr && !arg.empty())
+          {
+            stc->get_vi().command(arg);
+          }
+        }
+        handled = true;
+      }
 
       if (handled && wex::ex::get_macros().mode().is_playback())
       {